requirements.txt
azure-identity==1.23.0
azure-ai-projects==1.0.0b11
azure-ai-agents==1.1.0b2
azure-ai-inference[prompts]==1.0.0b9
fastapi==0.115.12
uvicorn==0.34.2
langchain==0.3.25
langchain-core==0.3.64
langchain-openai==0.3.21
langchain-azure-ai==0.1.4
langchain-mcp-adapters==0.1.7
langgraph==0.4.8
openai==1.85.0
pydantic==2.11.4
python-dotenv==1.1.0
requests==2.32.3
semantic-kernel==1.33.0
streamlit==1.45.1
fastmcp==2.7.1
# a2a-sdk==0.2.6
httpx==0.28.1
opentelemetry-api==1.31.1
azure.monitor.opentelemetry==1.6.10
mem0ai==0.1.104
azure-search-documents==11.5.2